+++ example/src/main/java/example/support/MessageManager.java
@@ -1,4 +1,4 @@
-package com.tagmycode.plugin.examples.support;
+package example.support;
 
 import com.tagmycode.plugin.GuiThread;
 import com.tagmycode.plugin.IMessageManager;