+++ src/main/java/com/tagmycode/plugin/gui/form/SnippetsTab.java
@@ -20,6 +20,7 @@ import java.awt.event.*;
 import static com.tagmycode.plugin.gui.GuiUtil.setPlaceholder;
 
 public class SnippetsTab extends AbstractGui implements IOnErrorCallback {
+    private final WelcomeView welcomeView;
     protected JButton editSnippetButton;
     protected JButton deleteSnippetButton;
     protected JButton copyButton;
@@ -41,16 +42,18 @@ public class SnippetsTab extends AbstractGui implements IOnErrorCallback {
     private ClipboardCopy clipboardCopy = new ClipboardCopy();
     private SnippetsTableModel model;
     private boolean networkingEnabled;
+    private FilterSnippetsTextField filterTextField;
 
     public SnippetsTab(final Framework framework) {
         this.framework = framework;
-        reset();
         initSnippetsJTable();
+        welcomeView = new WelcomeView(this);
 
         leftPane.add(snippetsTable.getMainComponent(), BorderLayout.CENTER);
         initFilterField();
         initToolBarButtons(framework);
         initPopupMenuForJTextComponents(getMainComponent());
+        reset();
     }
 
     private void initToolBarButtons(final Framework framework) {
@@ -143,7 +146,18 @@ public class SnippetsTab extends AbstractGui implements IOnErrorCallback {
             }
         });
 
+        addKeyStroke(this.jTable, KeyEvent.VK_F, new ActionListener() {
+            public void actionPerformed(ActionEvent e) {
+                filterTextField.requestFocus();
+            }
+        });
+
         initTablePopupMenu();
+    }
+
+    private void addKeyStroke(JComponent jComponent, int keyEvent, ActionListener actionListener) {
+        KeyStroke stroke = KeyStroke.getKeyStroke(keyEvent, Toolkit.getDefaultToolkit().getMenuShortcutKeyMask());
+        jComponent.registerKeyboardAction(actionListener, stroke, JComponent.WHEN_ANCESTOR_OF_FOCUSED_COMPONENT);
     }
 
     private void initTablePopupMenu() {
@@ -267,22 +281,28 @@ public class SnippetsTab extends AbstractGui implements IOnErrorCallback {
             public void valueChanged(ListSelectionEvent e) {
                 selectedRow = e.getFirstIndex();
                 if (!e.getValueIsAdjusting()) {
-                    snippetViewFormPane.removeAll();
-
                     Snippet snippet = snippetsTable.getSelectedSnippet();
-
-                    if (snippet != null) {
-                        JComponent snippetViewForm = new SnippetView(snippet).getMainComponent();
-                        snippetViewFormPane.add(snippetViewForm);
-                        enableButtonsForSnippet();
-                    } else {
-                        disableButtonsForSnippet();
-                    }
-                    snippetViewFormPane.revalidate();
-                    snippetViewFormPane.repaint();
+                    changeViewBasedOnSnippet(snippet);
                 }
             }
         };
+    }
+
+    private void changeViewBasedOnSnippet(Snippet snippet) {
+        snippetViewFormPane.removeAll();
+
+        if (snippet != null) {
+            JComponent snippetViewForm = new SnippetView(snippet).getMainComponent();
+            snippetViewFormPane.add(snippetViewForm);
+            enableButtonsForSnippet();
+        } else {
+            snippetViewFormPane.add(welcomeView.getMainComponent());
+            disableButtonsForSnippet();
+        }
+        snippetViewFormPane.revalidate();
+        snippetViewFormPane.repaint();
+        getMainComponent().revalidate();
+        getMainComponent().repaint();
     }
 
     private TableModelListener createTableModelListener() {
@@ -343,12 +363,9 @@ public class SnippetsTab extends AbstractGui implements IOnErrorCallback {
         return framework;
     }
 
-    public void reset() {
-        snippetViewFormPane.removeAll();
-    }
 
     public void initFilterField() {
-        FilterSnippetsTextField filterTextField = new FilterSnippetsTextField(framework, snippetsTable);
+        filterTextField = new FilterSnippetsTextField(framework, snippetsTable);
         filterTextField.setMinimumSize(new Dimension(200, 25));
         filterTextField.addKeyListener(new MoveUpDownFilterFieldKeyListener(jTable));
         setPlaceholder("Filter snippets", filterTextField);
@@ -366,5 +383,9 @@ public class SnippetsTab extends AbstractGui implements IOnErrorCallback {
     public void setNetworkingEnabled(boolean networkingEnabled) {
         setNetworkIcon(networkingEnabled);
         syncButton.setEnabled(networkingEnabled);
+    }
+
+    public void reset() {
+        changeViewBasedOnSnippet(null);
     }
 }