+++ src/main/java/com/tagmycode/plugin/gui/table/SnippetsTable.java
@@ -11,8 +11,6 @@ import javax.swing.table.JTableHeader;
 import javax.swing.table.TableColumn;
 import javax.swing.table.TableRowSorter;
 import java.awt.*;
-import java.awt.event.ActionEvent;
-import java.awt.event.KeyEvent;
 import java.util.ArrayList;
 
 public class SnippetsTable extends AbstractSnippetsListGui {
@@ -29,15 +27,12 @@ public class SnippetsTable extends AbstractSnippetsListGui {
         sorter = new TableRowSorter<>(model);
         table.setRowSorter(sorter);
         table.setAutoResizeMode(JTable.AUTO_RESIZE_LAST_COLUMN);
-        scrollPane = new JScrollPane(table);
-        scrollPane.setBorder(BorderFactory.createEmptyBorder());
         table.setIntercellSpacing(new Dimension(0, 0));
-
         table.setCellSelectionEnabled(false);
         table.setRowSelectionAllowed(true);
         table.setShowGrid(false);
-
-        createKeyBindings();
+        scrollPane = new JScrollPane(table);
+        scrollPane.setBorder(BorderFactory.createEmptyBorder());
 
         cellSelectionModel = table.getSelectionModel();
         cellSelectionModel.setSelectionMode(ListSelectionModel.SINGLE_SELECTION);
@@ -45,16 +40,6 @@ public class SnippetsTable extends AbstractSnippetsListGui {
         configureTableHeader();
 
         sortByColumn(SnippetsTableModel.MODIFIED);
-    }
-
-    private void createKeyBindings() {
-        table.getInputMap(JComponent.WHEN_ANCESTOR_OF_FOCUSED_COMPONENT).put(KeyStroke.getKeyStroke(KeyEvent.VK_ENTER, 0), "Enter");
-        table.getActionMap().put("Enter", new AbstractAction() {
-            @Override
-            public void actionPerformed(ActionEvent ae) {
-                //do something on JTable enter pressed
-            }
-        });
     }
 
     private void sortByColumn(int columnIndexToSort) {