+++ src/test/java/support/FakeVersion.java
@@ -1,0 +1,11 @@
+package support;
+
+import com.tagmycode.plugin.AbstractVersion;
+
+public class FakeVersion extends AbstractVersion {
+
+    @Override
+    public String getPluginVersion() {
+        return "FAKE";
+    }
+}